Update README.md
This is a Machine learning oriented Application for diabetes detection based on the parameters that the user enters. It predicts whether the diabetes is detected or not by learning from the parameters. The model is trained based on the PIMA Indian Diabetes dataset.
The PIMA Indian Diabetes dataset include attributes:
•	Number of times pregnant
•	Plasma glucose
•	Diastolic blood pressure (mm Hg)
•	Triceps skin fold thickness (mm)
•	2-Hour serum insulin (mu U/ml)
•	Body mass index (weight in kg/(height in m)^2)
•	Diabetes pedigree function
•	Age (years)
•	Class variable (0 or 1)
This model is developed in Python using its packages for Data analysis and Machine learning:
•	Pandas
•	ScikitLearn
•	Streamlit
•	Cv2